% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fliteMethods.R
\name{coef.flite}
\alias{coef.flite}
\title{Extract model coefficients method for objects of class \code{"flite"}}
\usage{
\method{coef}{flite}(object, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{"flite"}, a result of a
call to \code{\link{flite}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
A numeric vector of length 4 with names
  \code{c("p[u]", "sigma[u]", "xi", "theta")}.
}
\description{
\code{coef} method for class \code{"flite"}.
}
