% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite_methods.R
\name{coef.lite}
\alias{coef.lite}
\title{Extract model coefficients method for objects of class \code{"lite"}}
\usage{
\method{coef}{lite}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"lite"}, a result of a
call to \code{\link{flite}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
A numeric vector of length 4 with names
  \code{c("p[u]", "sigma[u]", "xi", "theta")}.
}
\description{
\code{coef} method for class \code{"lite"}.
}
