% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estfun.R
\name{estfun}
\alias{estfun}
\alias{estfun.Bernoulli}
\alias{estfun.GP}
\title{Functions for the \code{estfun} method}
\usage{
\method{estfun}{Bernoulli}(x, ...)

\method{estfun}{GP}(x, eps = 1e-05, m = 3, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{Further arguments.}

\item{eps, m}{These arguments control the estimation of the observed
information in \code{gpObsInfo} when the GP shape parameter \eqn{\xi} is
very close to zero.  In these cases, direct calculation is unreliable.
\code{eps} is a (small, positive) numeric scalar.  If the absolute value
of the input value of \eqn{\xi}, that is, \code{pars[2]}, is smaller than
\code{eps} then we approximate the \code{[2, 2]} element using a Taylor
series expansion in \eqn{\xi}, evaluated up to and including the
\code{m}th term.}
}
\description{
Functions to calculate contribution to the score vector from individual
observations for a fitted model object.
}
