% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flite.R
\name{flite}
\alias{flite}
\title{Frequentist inference for time series extremes}
\usage{
flite(data, u, cluster, ...)
}
\arguments{
\item{data}{A numeric vector or numeric matrix of raw data.  If \code{data}
  is a matrix then the log-likelihood is constructed as the sum of
  (independent) contributions from different columns. A common situation is
  where each column relates to a different year.

  If \code{data} contains missing values then \code{\link{split_by_NAs}} is
  used to divide the data further into sequences of non-missing values,
  stored in different columns in a matrix.  Again, the log-likelihood
  is constructed as a sum of contributions from different columns.}

\item{u}{A numeric scalar.  Extreme value threshold applied to data.}

\item{cluster}{Explain ... vector if data is a vector, matrix if data is a matrix}

\item{...}{Arguments to be passed to \code{\link[exdex]{kgaps}}, namely
\code{k} and/or \code{inc_cens}.}
}
\description{
Explain
}
\details{
Add details
}
\examples{
x <- exdex::cheeseboro
u <- quantile(x, probs = 0.9, na.rm = TRUE)
y <- flite(x, u)
}
