% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite_methods.R
\name{logLik.flite}
\alias{logLik.flite}
\title{Extract log-likelihood for objects of class \code{"flite"}}
\usage{
\method{logLik}{flite}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"flite"}, a result of a call to
\code{\link{flite}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
An object of class \code{"logLik"}: a numeric scalar with
value equal to the maximised log-likelihood.  This is the sum of
contributions from three fitted models, from a Bernoulli model for
occurrences of threshold exceedances, a generalised Pareto model for
threshold excesses and a \eqn{K}-gaps model for the extremal index.
The returned object also has attributes \code{nobs}, the numbers of
observations used in each of these model fits, and \code{"df"}, which is
equal to the number of total number of parameters estimated (4).
}
\description{
\code{logLik} method for class \code{"flite"}.
}
