% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikVector.R
\name{logLik.logLikVector}
\alias{logLik.logLikVector}
\title{Sum log-likelihood contributions from individual observations}
\usage{
\method{logLik}{logLikVector}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"logLikVector"} return from a
\code{logLikVector} method.}

\item{...}{Further arguments.}
}
\description{
S3 logLik method for logLikVector objects
}
