% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite_methods.R
\name{nobs.flite}
\alias{nobs.flite}
\title{Extract the number of observations from a fit for class \code{"flite"}}
\usage{
\method{nobs}{flite}(object, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{"flite"}, a result of a
call to \code{\link{flite}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}

\item{model}{One of \code{c("gp", "kgaps", "bernoulli")}.}
}
\value{
A numeric vector of length 3 with names
  \code{c("p[u]", "gp", "theta")}.
}
\description{
\code{nobs} method for class \code{"flite"}.
}
