% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{predict.blite}
\alias{predict.blite}
\title{Predictive inference for the largest value observed in N years.}
\usage{
\method{predict}{blite}(
  object,
  type = c("i", "p", "d", "q", "r"),
  x = NULL,
  x_num = 100,
  n_years = 100,
  ny = NULL,
  level = 95,
  hpd = FALSE,
  lower_tail = TRUE,
  log = FALSE,
  big_q = 1000,
  ...
)
}
\arguments{
\item{object}{An object of class "evpost", a result of a call to
\code{\link[revdbayes]{rpost}} or \code{\link[revdbayes]{rpost_rcpp}} with \code{model = "gev"},
\code{model = "os"}, \code{model = "pp"} or \code{model == "bingp"}.
Calling these functions after a call to \code{rpost} or \code{rpost_rcpp}
with \code{model == "gp"} will produce an error, because inferences about
the probability of threshold exceedance are required, in addition to the
distribution of threshold excesses. The model is stored in
\code{object$model}.}

\item{type}{A character vector.  Indicates which type of inference is
  required:
\itemize{
  \item "i" for predictive intervals,
  \item "p" for the predictive distribution function,
  \item "d" for the predictive density function,
  \item "q" for the predictive quantile function,
  \item "r" for random generation from the predictive distribution.
}}

\item{x}{A numeric vector or a matrix with \code{n_years} columns.
  The meaning of \code{x} depends on \code{type}.
  \itemize{
    \item{\code{type = "p"} or \code{type = "d"}:} \code{x} contains
      quantiles at which to evaluate the distribution or density function.

      If \code{object$model == "bingp"} then no element of \code{x} can be
      less than the threshold \code{object$thresh}.

      If \code{x} is not supplied then \code{n_year}-specific defaults are
      set: vectors of length \code{x_num} from the 0.1\% quantile to the
      99\% quantile, subject all values being greater than the threshold.

    \item{\code{type = "q"}:} \code{x} contains probabilities in (0,1)
      at which to evaluate the quantile function.  Any values outside
      (0, 1) will be removed without warning.

      If \code{object$model == "bingp"} then no element of \code{p} can
      correspond to a predictive quantile that is below the threshold,
      \code{object$thresh}.  That is, no element of \code{p} can be less
      than the value of \code{predict.evpost(object,}
      \code{type = "q", x = object$thresh)}.

      If \code{x} is not supplied then a default value of
      \code{c(0.025, 0.25, 0.5, 0.75, 0.975)} is used.
    \item{\code{type = "i"} or \code{type = "r"}:} \code{x} is not relevant.
  }}

\item{x_num}{A numeric scalar.  If \code{type = "p"} or \code{type = "d"}
and \code{x} is not supplied then \code{x_num} gives the number of values
in \code{x} for each value in \code{n_years}.}

\item{n_years}{A numeric vector. Values of N.}

\item{level}{A numeric vector of values in (0, 100).
Only relevant when \code{type = "i"}.
Levels of predictive intervals for the largest value observed in
N years, i.e. level\% predictive intervals are returned.}

\item{hpd}{A logical scalar.
  Only relevant when \code{type = "i"}.

  If \code{hpd = FALSE} then the interval is
  equi-tailed, with its limits produced by
  \code{predict.evpost(}\code{object, type ="q", x = p)},
  where \code{p = c((1-level/100)/2,} \code{(1+level/100)/2)}.

  If \code{hpd = TRUE} then, in addition to the equi-tailed interval,
  the shortest possible level\% interval is calculated.
  If the predictive distribution is unimodal then this
  is a highest predictive density (HPD) interval.}

\item{lower_tail}{A logical scalar.
Only relevant when \code{type = "p"} or \code{type = "q"}.
If TRUE (default), (output or input) probabilities are
\eqn{P[X \leq x]}{P[X <= x]}, otherwise, \eqn{P[X > x]}{P[X > x]}.}

\item{log}{A logical scalar.  Only relevant when \code{type = "d"}.
If TRUE the log-density is returned.}

\item{big_q}{A numeric scalar.  Only relevant when \code{type = "q"}.
An initial upper bound for the desired quantiles to be passed to
\code{\link[stats]{uniroot}} (its argument \code{upper}) in the
search for the predictive quantiles.  If this is not sufficiently large
then it is increased until it does provide an upper bound.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\description{
\code{predict} method for class "blite".  Performs predictive inference
about the largest value to be observed over a future time period of
N years.  Predictive inferences accounts for uncertainty in model
parameters and for uncertainty owing to the variability of future
observations.
}
\examples{
### Cheeseboro wind gusts

cdata <- exdex::cheeseboro
# Each column of the matrix cdata corresponds to data from a different year
# blite() sets cluster automatically to correspond to column (year)
cpost <- blite(cdata, u = 45, k = 3)
summary(cpost)
predict(cpost, ny = 31 * 24)$long
}
