% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite_methods.R
\name{print.lite}
\alias{print.lite}
\title{Print method for a \code{"lite"} object}
\usage{
\method{print}{lite}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{"lite"}, a result of a call to
\code{\link{flite}}.}

\item{digits}{The argument \code{digits} to \code{\link{print.default}}.}

\item{...}{Additional arguments.  None are used in this function.}
}
\value{
The argument \code{x}, invisibly, as for all
  \code{\link[base]{print}} methods.
}
\description{
\code{print} method for an object \code{x} of class \code{"lite"}.
}
\details{
Prints the original call to \code{\link{flite}}
  and the estimates of the parameters \eqn{p_u, \sigma_u, \xi, \theta}.
}
