% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentistReturnLevels.R
\name{print.returnLevel}
\alias{print.returnLevel}
\title{Print method for returnLevel object}
\usage{
\method{print}{returnLevel}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{c("returnLevel", "lite")}, a result of
a call to \code{\link{returnLevel}}.}

\item{digits}{The argument \code{digits} to \code{\link{print.default}}.}

\item{...}{Additional arguments.  None are used in this function.}
}
\value{
The argument \code{x}, invisibly, as for all
  \code{\link[base]{print}} methods.
}
\description{
\code{print} method for an objects of class \code{c("returnLevel", "lite")}.
}
\details{
Prints the call to \code{\link{returnLevel}} and the estimates
  and 100\code{x$level}\% confidence limits for the \code{x$m}-year
  return level.
}
\section{Examples}{

See \code{\link{returnLevel}}.
}

\seealso{
\code{\link{returnLevel}}.
}
