% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite_methods.R
\name{print.summary.flite}
\alias{print.summary.flite}
\title{Print method for objects of class \code{"summary.flite"}}
\usage{
\method{print}{summary.flite}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.flite", a result of a call to
\code{\link{summary.flite}}.}

\item{...}{Additional arguments passed on to \code{\link{print.default}}.}
}
\value{
Prints the numeric matrix returned from \code{\link{summary.flite}}.
}
\description{
\code{print} method for an object \code{x} of class \code{"summary.flite"}.
}
\section{Examples}{

See the examples in \code{\link{flite}}.
}

