% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentistReturnLevels.R
\name{print.summary.retlev}
\alias{print.summary.retlev}
\title{Print method for objects of class \code{"summary.retlev"}}
\usage{
\method{print}{summary.retlev}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.retlev", a result of a call to
\code{\link{summary.retlev}}.}

\item{...}{Additional arguments passed on to \code{\link{print.default}}.}
}
\value{
The argument \code{x}, invisibly, as for all
  \code{\link[base]{print}} methods.
}
\description{
\code{print} method for an object \code{x} of class \code{"summary.retlev"}.
}
\details{
Prints the call and the numeric matrix \code{x$matrix} returned from
  \code{\link{summary.retlev}}.
}
\section{Examples}{

See the examples in \code{\link{returnLevel}}.
}

\seealso{
\code{\link{returnLevel}} to perform inferences about return
  levels.
}
