% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fliteMethods.R
\name{summary.flite}
\alias{summary.flite}
\title{Summarising times series extreme fits}
\usage{
\method{summary}{flite}(object, adjust = TRUE, digits = max(3, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{An object inheriting from class \code{"flite"}, a result of a
call to \code{\link{flite}}.}

\item{adjust}{A logical scalar.  If \code{adjust = TRUE} then the
standard errors of \eqn{(p_u, \sigma_u, \xi)} are estimated via a sandwich
estimator of the variance-covariance matrix. See \code{\link{flite}}.
Otherwise, the inverse of the observed information matrix is used.}

\item{digits}{An integer. Passed to \code{\link[base:Round]{signif}} to
round the values in the summary.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
An object containing the original function call and a matrix of
  estimates and estimated standard errors with row names
  \code{c("p[u]", "sigma[u]", "xi", "theta")}.  The object is printed by
  \code{\link{print.summary.flite}}.
}
\description{
\code{summary} method for class \code{"flite"}
}
\section{Examples}{

See the examples in \code{\link{flite}}.
}

