% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentistReturnLevels.R
\name{summary.returnLevel}
\alias{summary.returnLevel}
\title{Summary method for a \code{"returnLevel"} object}
\usage{
\method{summary}{returnLevel}(object, digits, ...)
}
\arguments{
\item{object}{an object of class \code{c("returnLevel", "lite")}, a result of
a call to \code{\link{returnLevel}}.}

\item{digits}{An integer. Used for number formatting with
\code{\link[base:Round]{signif}}.  If \code{digits} is not specified
(i.e. \code{\link{missing}}) then \code{signif()} will not be called
(i.e. no rounding will be performed).}

\item{...}{Additional arguments.  None are used in this function.}
}
\value{
Returns a list containing the list element \code{object$call}
  and a numeric matrix \code{matrix} containing the MLE and estimated
  SE of the return level.
}
\description{
\code{summary} method for an objects of class \code{c("returnLevel", "lite")}.
}
\section{Examples}{

See \code{\link{returnLevel}}.
}

\seealso{
\code{\link{returnLevel}}.
}
