% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite_methods.R
\name{vcov.flite}
\alias{vcov.flite}
\title{Calculate the variance-covariance matrix for an object of class
\code{"flite"}}
\usage{
\method{vcov}{flite}(object, adjust = TRUE, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{"flite"}, a result of a
call to \code{\link{flite}}.}

\item{adjust}{A logical scalar.  If \code{adjust = TRUE} then the elements
of the variance-covariance matrix corresponding to
\eqn{(p_u, \sigma_u, \xi)} are estimated using a sandwich estimator.
See \code{\link{flite}}.  Otherwise, this matrix is the inverse of the
observed information matrix.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
A 4 by 4 numeric matrix with row and column names
  \code{c("p[u]", "sigma[u]", "xi", "theta")}.
}
\description{
\code{vcov} method for class \code{"flite"}.
}
