% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite_methods.R
\name{vcov.lite}
\alias{vcov.lite}
\title{Calculate the variance-covariance matrix for an object of class
\code{"lite"}}
\usage{
\method{vcov}{lite}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"lite"}, a result of a
call to \code{\link{flite}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
A 4 by 4 numeric matrix with row and column names
  \code{c("p[u]", "sigma[u]", "xi", "theta")}.
}
\description{
\code{vcov} method for class \code{"lite"}.
}
