% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernoulli.R
\name{Bernoulli}
\alias{Bernoulli}
\alias{fitBernoulli}
\alias{logLikVector.Bernoulli}
\alias{nobs.Bernoulli}
\alias{coef.Bernoulli}
\alias{vcov.Bernoulli}
\alias{logLik.Bernoulli}
\title{Inference for the Bernoulli distribution}
\usage{
fitBernoulli(data)

\method{logLikVector}{Bernoulli}(object, pars = NULL, ...)

\method{nobs}{Bernoulli}(object, ...)

\method{coef}{Bernoulli}(object, ...)

\method{vcov}{Bernoulli}(object, ...)

\method{logLik}{Bernoulli}(object, ...)
}
\arguments{
\item{data}{A numeric vector of outcomes from Bernoulli trials: 0 for a
 failure, 1 for a success.  Alternatively, a logical vector with FALSE
 for a failure and TRUE for a success. Missing values are removed using
\code{\link[stats:na.fail]{na.omit}}.}

\item{object}{A fitted model object returned from \code{fitBernoulli()}.}

\item{pars}{A numeric parameter vector of length 1 containing a value of
the Bernoulli success probability.}

\item{...}{Further arguments to be passed to the functions in the
sandwich package \code{\link[sandwich]{meat}} (if \code{cluster = NULL}),
or \code{\link[sandwich:vcovCL]{meatCL}} (if \code{cluster} is not
\code{NULL}).}
}
\value{
\code{fitBernoulli} returns an object of class \code{"bernoulli"}, a list
with components: \code{maxLogLik}, \code{mle}, \code{nobs}, \code{vcov},
\code{data}, \code{obs_data}, where \code{data} are the input data and
\code{obs_data} are the input data after any missing values have been
removed, using \code{\link[stats:na.fail]{na.omit}}.

\code{logLikVector.bernoulli} returns an object of class
\code{"logLikVector"}, a vector length \code{length(data)} containing the
likelihood contributions from the individual observations in \code{data}.
}
\description{
Functions involved in making inferences about the probability of success
in a Bernoulli distribution using maximum likelihood estimation.
}
\details{
\code{fitBernoulli}: fit a Bernoulli distribution using maximum likelihood
  estimation.

\code{logLikVector.bernoulli}: calculates contributions to an independence
log-likelihood based on the Bernoulli distribution.  The log-likelihood is
calculated up to an additive constant.

\code{nobs, coef, vcov} and \code{logLik} methods are provided.
}
\examples{
# Set up data
x <- exdex::newlyn
u <- quantile(x, probs = 0.9)
exc <- x > u

# Fit a Bernoulli distribution
fit <- fitBernoulli(exc)

# Calculate the log-likelihood at the MLE
res <- logLikVector(fit)

# The logLik method sums the individual log-likelihood contributions.
logLik(res)

# nobs, coef, vcov, logLik methods for objects returned from fitBernoulli()
nobs(fit)
coef(fit)
vcov(fit)
logLik(fit)
}
\seealso{
\code{\link[stats]{Binomial}}.  The Bernoulli distribution is the
  special case where \code{size = 1}.
}
