% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estfun.R
\name{estfun}
\alias{estfun}
\alias{estfun.Bernoulli}
\alias{estfun.GP}
\title{Functions for the \code{estfun} method}
\usage{
\method{estfun}{Bernoulli}(x, ...)

\method{estfun}{GP}(x, eps = 1e-05, m = 3, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{Further arguments.}
}
\description{
Functions to calculate contribution to the score vector from individual
observations for a fitted model object.
}
