% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fliteMethods.R
\name{fliteMethods}
\alias{fliteMethods}
\alias{coef.flite}
\alias{vcov.flite}
\alias{nobs.flite}
\alias{logLik.flite}
\alias{summary.flite}
\alias{print.summary.flite}
\title{Methods for objects of class \code{"flite"}}
\usage{
\method{coef}{flite}(object, ...)

\method{vcov}{flite}(object, adjust = TRUE, ...)

\method{nobs}{flite}(object, ...)

\method{logLik}{flite}(object, ...)

\method{summary}{flite}(object, adjust = TRUE, digits = max(3, getOption("digits") - 3L), ...)

\method{print}{summary.flite}(x, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{"flite"}, a result of a
call to \code{\link{flite}}.}

\item{...}{For \code{print.summary.flite}: additional arguments passed to
\code{\link{print.default}}.}

\item{adjust}{A logical scalar.  If \code{adjust = TRUE} then the
standard errors of \eqn{(p_u, \sigma_u, \xi)} are estimated via a sandwich
estimator of the variance-covariance matrix. See \code{\link{flite}}.
Otherwise, the inverse of the observed information matrix is used.}

\item{digits}{An integer. Passed to \code{\link[base:Round]{signif}} to
round the values in the summary.}

\item{x}{An object of class "summary.flite", a result of a call to
\code{\link{summary.flite}}.}
}
\value{
\code{coef.flite}: a numeric vector of length 4 with names
    \code{c("p[u]", "sigma[u]", "xi", "theta")}.

  \code{vcov.flite}: a 4 by 4 numeric matrix with row and column names
    \code{c("p[u]", "sigma[u]", "xi", "theta")}.

  \code{nobs.flite}: a numeric vector of length 3 with names
    \code{c("p[u]", "gp", "theta")}.

  \code{logLik.flite}: an object of class \code{"logLik"}: a numeric scalar
    with value equal to the maximised log-likelihood.  This is the sum of
    contributions from three fitted models, from a Bernoulli model for
    occurrences of threshold exceedances, a generalised Pareto model for
    threshold excesses and a \eqn{K}-gaps model for the extremal index.
    The returned object also has attributes \code{nobs}, the numbers of
    observations used in each of these model fits, and \code{"df"}, which is
    equal to the number of total number of parameters estimated (4).

  \code{summary.flite}: an object containing the original function call and
    a matrix of estimates and estimated standard errors with row names
    \code{c("p[u]", "sigma[u]", "xi", "theta")}.  The object is printed by
    \code{\link{print.summary.flite}}.

  \code{print.summary.flite}: prints the numeric matrix returned from
    \code{\link{summary.flite}}.
}
\description{
Methods for obects of class \code{"flite"} returned from
\code{\link{flite}}.
}
