% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite-internal.R
\name{lite-internal}
\alias{lite-internal}
\alias{adjust_object}
\alias{kgaps_loglik}
\alias{check_logLik_flite}
\alias{return_level_bingp}
\alias{bingp_rl_CI}
\alias{bingp_rl_prof}
\alias{bingp_rl_prof0}
\alias{box_cox_deriv}
\alias{make_ru_list}
\title{Internal lite functions}
\usage{
adjust_object(x, cluster = NULL, ...)

kgaps_loglik(theta, N0, N1, sum_qs, n_kgaps)

check_logLik_flite(object, ...)

return_level_bingp(x, m, level, ny, prof, inc, type, ny_given)

bingp_rl_CI(x, m, level, ny, type, u)

bingp_rl_prof(x, m, level, ny, inc, type, rl_sym, u)

bingp_rl_prof0(x, m, level, ny, inc, type, rl_sym, u)

box_cox_deriv(x, lambda = 1, lambda_tol = 1/50, poly_order = 3)

make_ru_list(model, trans, rotate, min_xi, max_xi)
}
\description{
Internal lite functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
