% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite-internal.R
\name{lite-internal}
\alias{lite-internal}
\alias{adjust_object}
\alias{kgaps_loglik}
\alias{check_logLik_lite}
\title{Internal lite functions}
\usage{
adjust_object(x, cluster = NULL, ...)

kgaps_loglik(theta, N0, N1, sum_qs, n_kgaps)

check_logLik_lite(object, ...)
}
\description{
Internal lite functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
