% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R
\docType{package}
\name{lite}
\alias{lite}
\title{lite: Likelihood-based Inference for Time Series Extremes}
\description{
Explain what the package does
}
\details{
Add details

The main function is \code{\link{flite}}, which performs frequentist
inference for time series extremes.

See \code{vignette("lite-vignette", package = "lite")} for an overview of the
package.
}
\references{
Northrop, P. J. and Christodoulides, C. (2022). exdex:
Estimation of the Extremal Index. R package version 1.1.1.
\url{https://CRAN.R-project.org/package=exdex/}.

Northrop, P. J. and Yin, C. (2022). lax: Loglikelihood Adjustment
for Extreme Value Models. R package version 1.2.0.
\url{https://CRAN.R-project.org/package=lax}

Northrop, P. J. (2020). revdbayes: Ratio-of-Uniforms Sampling
for Bayesian Extreme Value Analysis. R package version 1.3.9.
\url{https://paulnorthrop.github.io/revdbayes/}
}
