% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikVector.R
\name{logLikVector}
\alias{logLikVector}
\title{Evaluate log-likelihood contributions from specific observations}
\usage{
logLikVector(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Further arguments.}
}
\description{
Generic function for calculating log-likelihood contributions from
individual observations for a fitted model.
}
