% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikVector.R
\name{logLikVector}
\alias{logLikVector}
\alias{logLik.logLikVector}
\title{Functions for log-likelihoood contributions}
\usage{
logLikVector(x, ...)

\method{logLik}{logLikVector}(object, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{Further arguments.}

\item{object}{An object of class \code{"logLikVector"} returned from a
\code{logLikVector} method.}
}
\description{
Generic function to calculate log-likelihood contributions from
individual observations for a fitted model object.
}
