% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite_methods.R
\name{summary.lite}
\alias{summary.lite}
\title{Summarising times series extreme fits}
\usage{
\method{summary}{lite}(object, digits = max(3, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"lite"}, a result of a
call to \code{\link{flite}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
An object containing the original function call and a matrix of
  estimates and estimated standard errors with row names
  \code{c("p_u", "sigma_u", "xi", "theta")}.  The object is printed by
  \code{\link{print.summary.lite}}.
}
\description{
\code{summary} method for class \code{"lite"}
}
\section{Examples}{

See the examples in \code{\link{flite}}.
}

